pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "carver"
version = "1.0.0"
description = "A system for managing carver database and running pipelines"
readme = "README.md"
authors = [
    { name = "Venkata Pingali", email = "pingali@gmail.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
keywords = ["carver", "agents", "pipelines", "database"]
requires-python = ">=3.10"
dependencies = [
   "click>=8.0.0",
   "python-decouple>=3.8",
   "supabase>=2.0.0",
   "python-dateutil>=2.8.2",
   "tabulate>=0.9.0",
   "feedparser>=6.0.0",
   "requests>=2.31.0",
   "beautifulsoup4>=4.12.0",
   "google-api-python-client==2.153.0",
   "pytube",
   "youtube_transcript_api",
   "isodate==0.7.0",
   "lxml>=4.9.0",  # For better XML parsing
   "langdetect>=1.0.9",  # For language detection
   "bs4"
]

[project.urls]
Homepage = "https://github.com/pingali/carver"
Documentation = "https://github.com/pingali/carver#readme"
Repository = "https://github.com/pingali/carver.git"
Issues = "https://github.com/pingali/carver/issues"

[project.scripts]
carver-run = "carver.cli:cli"

[tool.setuptools]
packages = ["carver"]

[tool.setuptools.package-data]
carver = ["py.typed"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=carver --cov-report=term-missing"
testpaths = [
    "tests",
]
